requirements.txt
python-docx>=1.1.0
pillow>=10.0.0
